# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Zord backend server.",
        "title": "Swagger Zord API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/dummy": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "List Dummy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "valid int",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "value example: eql|lik,value",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "value example: lik,value",
                        "name": "email",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "tenant name",
                        "name": "Tenant",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_LIST.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Create Dummy",
                "parameters": [
                    {
                        "description": "body model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_CREATE.Data"
                        }
                    },
                    {
                        "type": "string",
                        "description": "tenant name",
                        "name": "Tenant",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_CREATE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        },
        "/dummy/{dummy_id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Get a Dummy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dummy ID",
                        "name": "dummy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tenant name",
                        "name": "Tenant",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_GET.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Edit Dummy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dummy ID",
                        "name": "dummy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "body model",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_EDIT.Data"
                        }
                    },
                    {
                        "type": "string",
                        "description": "tenant name",
                        "name": "Tenant",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_EDIT.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dummy"
                ],
                "summary": "Delete Dummy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dummy ID",
                        "name": "dummy_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "tenant name",
                        "name": "Tenant",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_DELETE.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/go-skeleton_internal_application_services.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "go-skeleton_internal_application_domain_dummy.Dummy": {
            "type": "object",
            "properties": {
                "DummyName": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                },
                "ID": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services.Error": {
            "type": "object",
            "properties": {
                "Error": {
                    "type": "string"
                },
                "Message": {},
                "Status": {
                    "type": "integer"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_CREATE.Data": {
            "type": "object",
            "required": [
                "DummyName",
                "Email"
            ],
            "properties": {
                "DummyName": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 3
                },
                "Email": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_CREATE.Response": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/go-skeleton_internal_application_domain_dummy.Dummy"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_DELETE.Response": {
            "type": "object",
            "properties": {
                "Message": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_EDIT.Data": {
            "type": "object",
            "required": [
                "DummyName",
                "Email"
            ],
            "properties": {
                "DummyName": {
                    "type": "string"
                },
                "Email": {
                    "type": "string"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_EDIT.Response": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/go-skeleton_internal_application_services_dummy_EDIT.Data"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_GET.Response": {
            "type": "object",
            "properties": {
                "Data": {
                    "$ref": "#/definitions/go-skeleton_internal_application_domain_dummy.Dummy"
                }
            }
        },
        "go-skeleton_internal_application_services_dummy_LIST.Response": {
            "type": "object",
            "properties": {
                "CurrentPage": {
                    "type": "integer"
                },
                "Data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/go-skeleton_internal_application_domain_dummy.Dummy"
                    }
                },
                "TotalPages": {
                    "type": "integer"
                }
            }
        }
    }
}